+++ pkg/docker-exec-ws/shellclient.go
@@ -10,7 +10,7 @@ import (
 	"time"
 
 	"github.com/gorilla/websocket"
-	"github.com/taskcluster/taskcluster-cli/external/pkg/iopipes"
+	"github.com/taskcluster/taskcluster-cli/pkg/iopipes"
 	"github.com/taskcluster/taskcluster-worker/plugins/interactive/shellconsts"
 	"github.com/taskcluster/taskcluster-worker/runtime/atomics"
 )
